# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Elite Starx earning platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/settings:
    get:
      operationId: adminGetSettings
      tags: [admin]
      summary: Get platform settings (admin)
      responses:
        "200":
          description: List of settings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Setting"
    patch:
      operationId: adminUpdateSetting
      tags: [admin]
      summary: Update a platform setting (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingBody"
      responses:
        "200":
          description: Setting updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /admin/users/{id}/block:
    patch:
      operationId: adminBlockUser
      tags: [admin]
      summary: Block or unblock a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlockUserBody"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/users/{id}/balance:
    patch:
      operationId: adminAdjustBalance
      tags: [admin]
      summary: Adjust user balance (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdjustBalanceBody"
      responses:
        "200":
          description: Balance adjusted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /admin/deposits/{id}/approve:
    patch:
      operationId: adminApproveDeposit
      tags: [admin]
      summary: Approve or reject a deposit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveRequestBody"
      responses:
        "200":
          description: Deposit updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /admin/withdrawals/{id}/approve:
    patch:
      operationId: adminApproveWithdrawal
      tags: [admin]
      summary: Approve or reject a withdrawal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveRequestBody"
      responses:
        "200":
          description: Withdrawal updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    BlockUserBody:
      type: object
      properties:
        blocked:
          type: boolean
      required:
        - blocked
    AdjustBalanceBody:
      type: object
      properties:
        amount:
          type: number
        reason:
          type: string
      required:
        - amount
        - reason
...
    ApproveRequestBody:
      type: object
      properties:
        action:
          type: string
          enum: [approve, reject]
        notes:
          type: ["string", "null"]
      required:
        - action
...
    Setting:
      type: object
      properties:
        key:
          type: string
        value:
          type: string
        updatedAt:
          type: string
      required:
        - key
        - value
        - updatedAt
    UpdateSettingBody:
      type: object
      properties:
        key:
          type: string
        value:
          type: string
      required:
        - key
        - value