# lib/api-zod/src/generated/types/successResponse.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Api
 * Elite Starx earning platform API
 * OpenAPI spec version: 0.1.0
 */

export interface SuccessResponse {
  success: boolean;
  message?: string;
}